Code refactoring using patterns (#32)
* Completely implement singleton for CManager

* Check if duplicate manager when adding to manager list

* Create manager class documentation

* Fix message printing

* Make managers created when program starts
Add Reload method to the server instance class
Add user database proxy to measure methods execution time

* Add documentation for command classes

* Check for command duplicate when adding to list

* Remove temporary DB_PROXY define (should be defined when using SERVER_DB_PROXY option)

* Add missing EOL in CManager::AddManager log

* Don't print this message if client leaves

* Separate loading and parsing event quests in CQuestManager
Add tasks and conditions in constructors, not in parse method

* Add manager tests

* Add ModeEvent.csv metadata
Send ban list only if it exists in database

* Replace int with size_t in CSendPacket::WriteData method

* Move server command definitions to separate file

* Fix a problem where user info and inventory not showing on client
Add new room settings variable
Add sendpacket command

* Add missing packet ids and rename existing ones

* Add unknown hostserver packet ids and add transfer function

* Update metadata to latest

* Don't change FD_SETSIZE define value to avoid memory corruption and crash in random places

* Rewrite server events
Use target_compile_definitions instead of add_definitions in cmakelists

* Add test for events

* Refactoring logging code:
- separate logging to file and console
- composite logging

* Some fixes in tests

* Template method for CBaseLogger

* Move logger test classes inside test case

* Delete readme.txt

* Add test for commands

---------

+++ src/public/net/sendpacket.h
@@ -23,7 +23,7 @@ public:
 	void WriteUInt64(unsigned long long number, bool littleEndian = true);
 	void WriteString(const std::string& str);
 	void WriteWString(const std::wstring& str);
-	void WriteData(void* data, int len);
+	void WriteData(void* data, size_t len);
 	void WriteArray(const std::vector<unsigned char>& arr);
 	void SetWriteOffset(int offset);
 	void SetOverride(bool override);
